move root reducer into its own file
